\name{ComputeTE-package}
\alias{ComputeTE-package}
\alias{ComputeTE}
\docType{package}
\title{
\packageTitle{ComputeTE}
}
\description{
\packageDescription{ComputeTE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ComputeTE}
\packageIndices{ComputeTE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ComputeTE}

Maintainer: \packageMaintainer{ComputeTE}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}