% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ComputeTE}
\alias{ComputeTE}
\alias{te}
\alias{transfer-entropy}
\alias{transferentropy}
\title{Estimate Transfer Entropy.}
\usage{
ComputeTE(X, Y, embedding, k, method = "MI_diff", epsDist = -1)
}
\arguments{
\item{X}{Numeric vector, Transfer Entropy is calculated to random process X}

\item{Y}{Numeric vector, Transfer Entropy is calculated from random process Y}

\item{embedding}{Numeric, The embedding dimension. It should be positive}

\item{k}{Numeric, The k'th neighbor. It should be positive}

\item{method}{String, The method to be used to estimate TE from ("MI_dif","Direct","Correlation")}

\item{epsDistace}{Numeric, The distance used for measuring TE in Correlation method, by default it is the average distance calculated in XKY}
}
\value{
Numeric, The estimated transfer entropy
}
\description{
\code{ComputeTE} return estimation of transfer entropy between two random process given a method.
}
\details{
A function to calculate Transfer entropy from a random process \code{Y} to
the random process \code{X} with specified \code{embedding} dimension and \code{k}'th neighbor for
k nearest neibor search used in the estimation given a \code{method} from provided methods in this package.
TODO: fix details (talk about adding noise somewhere)
}
\examples{
X <- rep(0,10000+1)
Y <- rep(0,10000+1)
for(i in 1:10000){
  Y[i+1] <- 0.5*Y[i] + rnorm(1)
  X[i+1] <- 0.6*X[i] + 0.5*Y[i] + rnorm(1)
}
ComputeTE(X,Y,3,1,"MI_diff")
ComputeTE(Y,X,3,1,"MI_diff")
ComputeTE(X,Y,1,1,"Correlation")
ComputeTE(Y,X,1,1,"Correlation")
ComputeTE(X,Y,3,1,"Correlation",0.5)
}
\seealso{
TODO arbitrary
}

