\name{TransferEntropy-package}
\alias{TransferEntropy-package}
\docType{package}
\title{
\packageTitle{TransferEntropy}
}
\description{
\packageDescription{TransferEntropy}
The main function \code{computeTE} calculate transfer entropy from one to another random process given a method from mentioned methods. 
}

\author{
\packageAuthor{TransferEntropy}
Maintainer: \packageMaintainer{TransferEntropy}
}

\references{
Estimating mutual information. Alexander Kraskov, Harald Stoegbauer, Peter Grassberger. 2004
A Recipe for the Estimation of Information Flow in a Dynamical System. Deniz Gencaga, Kevin H. Knuth, William B. Rossow. 2015.
}