\name{TransferEntropy-package}
\alias{TransferEntropy-package}
\docType{package}
\title{
\packageTitle{TransferEntropy}
}
\description{
\packageDescription{TransferEntropy}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TransferEntropy}
\packageIndices{TransferEntropy}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TransferEntropy}

Maintainer: \packageMaintainer{TransferEntropy}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}